Test 1

+++ code/renderergl1/tr_local.h
@@ -33,7 +33,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "../renderercommon/qgl.h"
 
 #define GL_INDEX_TYPE		GL_UNSIGNED_INT
-typedef unsigned int glIndex_t;
+typedef uint16_t glIndex_t;
 
 // 14 bits
 // can't be increased without changing bit packing for drawsurfs
